경매일정 상세페이지 수정, 삭제

+++ artauction/src/components/Auction/AuctionScheduleDetail.js
@@ -1,20 +1,24 @@
 import { Navigate, useNavigate, useParams } from "react-router";
 import Jumbotron from "../Jumbotron";
-import { useCallback, useEffect, useState } from "react";
+import { useCallback, useEffect, useRef, useState } from "react";
 import axios from "axios";
+import { Modal } from "bootstrap";
+import moment from "moment";
+import "moment/locale/ko";  
+moment.locale("ko");  
 
 
 const AuctionScheduleDetail = ()=>{
      //parameter
      const {auctionScheduleNo} = useParams();
-
+     
      //navigator
      const navigate = useNavigate();
 
      //state
      const [auctionSchedule, setAuctionSchedule] = useState({});
 
-     const [input, setInput] = useState({
+     const [target, setTarget] = useState({    //수정
         auctionScheduleNo : "",
         auctionScheduleTitle : "",
         auctionScheduleStartDate : "",
@@ -34,26 +38,58 @@ const AuctionScheduleDetail = ()=>{
             setAuctionSchedule(resp.data);
     }, [auctionSchedule, auctionScheduleNo]);
 
-    const changeInput = useCallback(e=>{
-        setInput({
-            ...input,
+    //수정
+    const changeTarget = useCallback(e=>{
+        setTarget({
+            ...target,
             [e.target.name] : e.target.value
-        })
-    }, [input]);
+        });
+    }, [target]);
 
-    const editAuctionSchedule = useCallback(async ()=>{
-        await axios.put("http://localhost:8080/auctionSchedule/", input);
-        changeInput();
-        loadAuctionSchedule();
-    }, [auctionSchedule, input]);
+    const clearTarget = useCallback(e=>{
+        setTarget({
+            auctionScheduleNo : "",
+            auctionScheduleTitle : "",
+            auctionScheduleStartDate : "",
+            auctionScheduleEndDate : "",
+            auctionScheduleState : "",
+            auctionScheduleNotice : "",
+        })
+    }, [target])
+
+    const saveTarget = useCallback(async ()=>{
+        const resp = await axios.put("http://localhost:8080/auctionSchedule/", target);
+        loadAuctionSchedule(resp.data);
+        closeChangeModal();
+    }, [target]);
+
+
+    // 수정 모달
+    const changeModal = useRef();
+
+    const openChangeModal = useCallback((auctionSchedule)=>{
+        const tag = Modal.getOrCreateInstance(changeModal.current);
+        tag.show();
+        setTarget({...auctionSchedule});  
+    }, [changeModal]);
+
+    const closeChangeModal = useCallback(()=>{
+        const tag = Modal.getInstance(changeModal.current);
+        tag.hide();
+        clearTarget();
+    }, [changeModal]);
+
 
 
     const deleteAuctionSchedule = useCallback(async ()=>{
+        const choice = window.confirm("정말 삭제하시겠습니까?");
+        if(choice === false) return;
+
         await axios.delete("http://localhost:8080/auctionSchedule/"+auctionScheduleNo);
         navigate("/auctionschedule");
     }, [auctionSchedule, auctionScheduleNo]);
 
-    
+
     //view
     return (<>
         <Jumbotron title={auctionScheduleNo +"번 경매 일정 상세정보"}/>
@@ -62,7 +98,7 @@ const AuctionScheduleDetail = ()=>{
 
             <div className="row mt-4 text-center">
                 <div className="col">
-                    <img src="https://placehold.co/300" class="img-thumbnail" alt=""/>           
+                    <img src="https://placehold.co/300" className="img-thumbnail" alt=""/>           
                 </div>
             </div>
 
@@ -86,7 +122,7 @@ const AuctionScheduleDetail = ()=>{
                     시작일
                 </div>
                 <div className="col-sm-9">
-                    {auctionSchedule.auctionScheduleStartDate}
+                    {moment(auctionSchedule.auctionScheduleStartDate).format("yyyy/MM/DD (dd) a hh:mm")}
                 </div>
             </div>
 
@@ -95,7 +131,7 @@ const AuctionScheduleDetail = ()=>{
                     종료일
                 </div>
                 <div className="col-sm-9">
-                    {auctionSchedule.auctionScheduleEndDate}
+                    {moment(auctionSchedule.auctionScheduleEndDate).format("yyyy/MM/DD (dd) a hh:mm")}
                 </div>
             </div>
             
@@ -115,10 +151,81 @@ const AuctionScheduleDetail = ()=>{
             <div className="col text-end">
                 <button className="btn btn-secondary ms-2" 
                                 onClick={e=>navigate("/auctionschedule")}>목록보기</button>
-                    <button className="btn btn-warning ms-2" 
-                                onClick={editAuctionSchedule}>수정하기</button>
-                    <button className="btn btn-danger ms-2" 
-                                onClick={deleteAuctionSchedule}>삭제하기</button>
+                <button className="btn btn-warning ms-2" 
+                             onClick={e=>openChangeModal(auctionSchedule)}>수정하기</button>
+                <button className="btn btn-danger ms-2" 
+                            onClick={e=>deleteAuctionSchedule(auctionSchedule)}>삭제하기</button>
+            </div>
+        </div>
+
+
+        {/* 경매 일정 수정 모달 */}
+        <div className="modal fade" tabIndex="-1" 
+                    ref={changeModal} data-bs-backdrop="static">
+            <div className="modal-dialog">
+                <div className="modal-content">
+
+                    <div className="modal-header">
+                        <h5 className="modal-title">
+                            경매 일정 수정
+                         </h5>
+                        <button type="button" className="btn-close" 
+                                data-bs-dismiss="modal" aria-label="Close"
+                                onClick={closeChangeModal}>
+                        <span aria-hidden="true"></span>
+                        </button>
+                     </div>
+
+                    <div className="modal-body">
+                        <div className="row mt-4">
+                            <div className="col">
+                                 <label>경매 일정명</label>
+                                <input type="text" className="form-control mb-4" 
+                                        name="auctionScheduleTitle" 
+                                        value={target.auctionScheduleTitle} 
+                                        onChange={changeTarget}/>
+                                            
+                                <label>일정 시작일</label>
+                                <input type="datetime-local" className="form-control mb-4" 
+                                        name="auctionScheduleStartDate" 
+                                        value={moment(target.auctionScheduleStartDate).format("yyyy-MM-DD a hh:mm")} 
+                                        onChange={changeTarget}/>
+
+                                <label>일정 종료일</label>
+                                <input type="datetime-local" className="form-control mb-4" 
+                                        name="auctionScheduleEndDate" 
+                                        value={moment(target.auctionScheduleEndDate).format("yyyy-MM-DD a hh:mm")} 
+                                        onChange={changeTarget}/>
+                                            
+                                <label>일정 상태</label>
+                                <select className="form-select mb-4" name="auctionScheduleState" 
+                                        value={target.auctionScheduleState} onChange={changeTarget}>
+                                    <option value="">선택하세요</option>
+                                    <option>예정경매</option>
+                                    <option>진행경매</option>
+                                    <option>종료경매</option>
+                                </select>
+
+                                <label>안내사항</label>
+                                <textarea type="text" className="form-control mb-4" 
+                                            name="auctionScheduleNotice" 
+                                            value={target.auctionScheduleNotice} 
+                                            onChange={changeTarget}/>
+
+                                <label>이미지</label>
+
+                            </div>
+                        </div>
+                    </div>
+
+                    <div className="modal-footer">
+                        <button type="button" className="btn btn-secondary btn-manual-close" 
+                                    onClick={closeChangeModal}>취소</button>
+                        <button type="button" className="btn btn-success"
+                                    onClick={saveTarget}>수정</button>       
+                    </div>
+
+                </div>
             </div>
         </div>
